+++ app/src/androidTest/java/woowacourse/movie/uiTest/ReservationActivityTest.kt
@@ -15,8 +15,8 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.TestName
 import woowacourse.movie.R
-import woowacourse.movie.domain.Date
 import woowacourse.movie.domain.Movie
+import woowacourse.movie.domain.movietime.Date
 import woowacourse.movie.uiTest.fixture.fakeContext
 import woowacourse.movie.view.reservation.detail.ReservationActivity
 import java.time.LocalDate
