+++ app/src/main/java/woowacourse/movie/domain/Movie.kt
@@ -2,6 +2,7 @@ package woowacourse.movie.domain
 
 import androidx.annotation.DrawableRes
 import woowacourse.movie.R
+import woowacourse.movie.domain.movietime.Date
 import java.io.Serializable
 import java.time.LocalDate
