+++ app/src/main/java/woowacourse/movie/domain/movieseat/Seats.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.movieseat
 
 import java.io.Serializable
