+++ app/src/main/java/woowacourse/movie/domain/movietime/MovieSchedule.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.movietime
 
 import java.time.LocalDate
