+++ app/src/main/java/woowacourse/movie/domain/movietime/ScreeningTime.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.movietime
 
 import java.time.DayOfWeek
 import java.time.LocalDate
