+++ app/src/main/java/woowacourse/movie/view/reservation/result/ReservationCompleteActivity.kt
@@ -2,7 +2,6 @@ package woowacourse.movie.view.reservation.result
 
 import android.content.Context
 import android.content.Intent
-import android.icu.text.DecimalFormat
 import android.os.Build
 import android.os.Bundle
 import android.widget.TextView
@@ -11,8 +10,8 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import woowacourse.movie.R
-import woowacourse.movie.domain.Seats
 import woowacourse.movie.domain.Ticket
+import woowacourse.movie.domain.movieseat.Seats
 import woowacourse.movie.view.dialog.DialogFactory
 import woowacourse.movie.view.reservation.detail.ReservationActivity
 import java.time.format.DateTimeFormatter
