+++ app/src/main/java/woowacourse/movie/view/reservation/result/ReservationCompleteContract.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.view.reservation.result
 
-import woowacourse.movie.domain.Seats
 import woowacourse.movie.domain.Ticket
+import woowacourse.movie.domain.movieseat.Seats
 
 class ReservationCompleteContract {
     interface Presenter {
