+++ app/src/main/java/woowacourse/movie/view/reservation/seat/ReservationSeatPresenter.kt
@@ -1,9 +1,9 @@
 package woowacourse.movie.view.reservation.seat
 
-import woowacourse.movie.domain.Position
-import woowacourse.movie.domain.Seat
-import woowacourse.movie.domain.Seats
 import woowacourse.movie.domain.Ticket
+import woowacourse.movie.domain.movieseat.Position
+import woowacourse.movie.domain.movieseat.Seat
+import woowacourse.movie.domain.movieseat.Seats
 
 class ReservationSeatPresenter(
     val view: ReservationSeatContract.View,
