+++ app/src/test/java/woowacourse/movie/domain/ScreeningTimeTest.kt
@@ -3,6 +3,7 @@ package woowacourse.movie.domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import woowacourse.movie.domain.movietime.ScreeningTime
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
